change signin directory rayer
